Move partial date utils to dedicated file
