+++ apps/website/src/pages/about/alveus.tsx
@@ -2,9 +2,10 @@ import { type NextPage } from "next";
 import Image from "next/image";
 import { type ReactNode } from "react";
 
-import type { PartialDateString } from "@alveusgg/data/build/types";
-
-import { formatPartialDateString } from "@/utils/datetime";
+import {
+  type PartialDateString,
+  formatPartialDateString,
+} from "@/utils/datetime-partial";
 
 import Box from "@/components/content/Box";
 import Button from "@/components/content/Button";
