+++ apps/website/src/pages/ambassadors/index.tsx
@@ -17,7 +17,7 @@ import { classes } from "@/utils/classes";
 import {
   parsePartialDateString,
   sortPartialDateString,
-} from "@/utils/datetime";
+} from "@/utils/datetime-partial";
 import { typeSafeObjectEntries } from "@/utils/helpers";
 import { convertToSlug } from "@/utils/slugs";
 import { camelToKebab } from "@/utils/string-case";
